+++ app/components/UI/PaymentRequest/index.js
@@ -54,7 +54,7 @@ import {
   REQUEST_AMOUNT_INPUT,
   REQUEST_SEARCH_ASSET_INPUT,
   REQUEST_SEARCH_SCREEN,
-} from '../../../../wdio/screen-objects/testIDs/Screens/RecieveToken.testIds';
+} from '../../../../wdio/screen-objects/testIDs/Screens/RequestToken.testIds';
 
 const KEYBOARD_OFFSET = 120;
 const createStyles = (colors) =>
