+++ wdio/step-definitions/request-token.steps.js
@@ -1,25 +1,24 @@
 /* global driver */
 import { Then, When } from '@wdio/cucumber-framework';
+
 import RequestTokenScreen from '../screen-objects/RequestTokenScreen';
+import SendLinkScreen from '../screen-objects/SendLinkScreen';
 
 Then(/^I tap on the close payment request icon/, async () => {
   const timeout = 1000;
   await driver.pause(timeout);
-  await RequestTokenScreen.closePaymentRequest();
+  await SendLinkScreen.closePaymentRequest();
 });
 
 When(/^I type "([^"]*)?" into the Request Amount field/, async (amount) => {
   await RequestTokenScreen.typeAmountInRequest(amount);
 });
 
-Then(
-  /^I type "([^"]*)?" in the Search Assets field/,
-  async (seaarchRequest) => {
-    const timeout = 1000;
-    await driver.pause(timeout);
-    await RequestTokenScreen.inputSearchRequestField(seaarchRequest);
-  },
-);
+Then(/^I type "([^"]*)?" in the Search Assets field/, async (searchRequest) => {
+  const timeout = 1000;
+  await driver.pause(timeout);
+  await RequestTokenScreen.inputSearchRequestField(searchRequest);
+});
 
 Then(/^I tap to navigate back from request view/, async () => {
   const timeout = 1000;
@@ -32,6 +31,6 @@ Then(/^I am taken back to the Request Search view/, async () => {
 });
 
 Then(/^I close the request screen/, async () => {
-  await RequestTokenScreen.closeQRPayment();
-  await RequestTokenScreen.closePaymentRequest();
+  await SendLinkScreen.closeQRPayment();
+  await SendLinkScreen.closePaymentRequest();
 });
